refactor: move the statistics to the plugin's settings
