+++ package.json
@@ -20,17 +20,17 @@
     "author": "Stephen Mwangi",
     "license": "MIT",
     "devDependencies": {
-        "@eslint/js": "^9.11.0",
+        "@eslint/js": "^9.11.1",
         "@types/eslint__js": "^8.42.3",
         "@types/jest": "^29.5.13",
-        "@types/node": "^22.5.5",
+        "@types/node": "^22.7.1",
         "@types/vhtml": "^2.2.9",
-        "@typescript-eslint/eslint-plugin": "^8.6.0",
-        "@typescript-eslint/parser": "^8.6.0",
+        "@typescript-eslint/eslint-plugin": "^8.7.0",
+        "@typescript-eslint/parser": "^8.7.0",
         "auto-changelog": "^2.5.0",
         "builtin-modules": "^4.0.0",
         "esbuild": "^0.23.1",
-        "eslint": "^9.11.0",
+        "eslint": "^9.11.1",
         "eslint-plugin-simple-import-sort": "^12.1.1",
         "eslint-plugin-unicorn": "^55.0.0",
         "jest": "^29.7.0",
@@ -42,7 +42,7 @@
         "ts-jest": "^29.2.5",
         "tslib": "^2.7.0",
         "typescript": "~5.5.4",
-        "typescript-eslint": "^8.6.0",
+        "typescript-eslint": "^8.7.0",
         "vhtml": "^2.2.0"
     },
     "dependencies": {
