+++ src/lang/locale/cz.ts
@@ -50,7 +50,7 @@ export default {
     YEARS_STR_IVL_MOBILE: "${interval}r",
 
     // settings.ts
-    SETTINGS_HEADER: "Spaced Repetition - Nastavení",
+    SETTINGS_HEADER: "Spaced Repetition",
     GROUP_TAGS_FOLDERS: "Tags & Folders",
     GROUP_FLASHCARD_REVIEW: "Flashcard Review",
     GROUP_FLASHCARD_SEPARATORS: "Flashcard Separators",
