+++ src/lang/locale/es.ts
@@ -50,7 +50,7 @@ export default {
     YEARS_STR_IVL_MOBILE: "${interval}a",
 
     // settings.ts
-    SETTINGS_HEADER: "Extensión de Repetición Espaciada - Ajustes",
+    SETTINGS_HEADER: "Extensión de Repetición Espaciada",
     GROUP_TAGS_FOLDERS: "Tags & Folders",
     GROUP_FLASHCARD_REVIEW: "Flashcard Review",
     GROUP_FLASHCARD_SEPARATORS: "Flashcard Separators",
