+++ src/lang/locale/fr.ts
@@ -51,7 +51,7 @@ export default {
     YEARS_STR_IVL_MOBILE: "${interval}a",
 
     // settings.ts
-    SETTINGS_HEADER: "Spaced Repetition - Paramètres",
+    SETTINGS_HEADER: "Spaced Repetition",
     GROUP_TAGS_FOLDERS: "Tags & Dossiers",
     GROUP_FLASHCARD_REVIEW: "Apprentissage des flashcards",
     GROUP_FLASHCARD_SEPARATORS: "Séparateurs de flashcards",
