+++ src/lang/locale/ru.ts
@@ -59,7 +59,7 @@ export default {
     YEARS_STR_IVL_MOBILE: "${interval}г.",
 
     // settings.ts
-    SETTINGS_HEADER: "Плагин Spaced Repetition - Настройки",
+    SETTINGS_HEADER: "Spaced Repetition",
     GROUP_TAGS_FOLDERS: "Tags & Folders",
     GROUP_FLASHCARD_REVIEW: "Flashcard Review",
     GROUP_FLASHCARD_SEPARATORS: "Flashcard Separators",
