+++ src/lang/locale/tr.ts
@@ -50,7 +50,7 @@ export default {
     YEARS_STR_IVL_MOBILE: "${interval}y",
 
     // settings.ts
-    SETTINGS_HEADER: "Aralıklı Tekrar - Ayarlar",
+    SETTINGS_HEADER: "Aralıklı Tekrar",
     GROUP_TAGS_FOLDERS: "Etiketler ve Klasörler",
     GROUP_FLASHCARD_REVIEW: "Flash Kartları Gözden Geçirme",
     GROUP_FLASHCARD_SEPARATORS: "Flash Kart Ayırıcıları",
